pyproject.toml
[project]
name = "governmentreporter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "chromadb>=1.0.15",
    "feedparser>=6.0.11",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "mcp[cli]>=1.12.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "tiktoken>=0.9.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "mypy>=1.17.0",
    "pytest>=8.4.1",
]
